Simplify crates more
