+++ harbor-ui/src/bridge.rs
@@ -6,9 +6,9 @@ use fedimint_core::config::FederationId;
 use fedimint_core::invite_code::InviteCode;
 use fedimint_core::Amount;
 use fedimint_ln_common::lightning_invoice::Bolt11Invoice;
-use harbor_client::core::HarborCore;
 use harbor_client::db::{check_password, setup_db, DBConnection};
 use harbor_client::fedimint_client::{FederationInviteOrId, FedimintClient};
+use harbor_client::HarborCore;
 use harbor_client::{CoreUIMsg, CoreUIMsgPacket, UICoreMsg, UICoreMsgPacket};
 use iced::futures::channel::mpsc::Sender;
 use iced::futures::{SinkExt, Stream, StreamExt};
